refactor: さらなるディレクトリ構成の整理
