+++ src/components/templates/CalcStatsTemplate.tsx
@@ -18,10 +18,10 @@ import { EffortValueField } from '@/components/organisms/EffortValueField'
 import { IndividualValueField } from '@/components/organisms/IndividualValueField'
 import { RealNumberField } from '@/components/organisms/RealNumberField'
 import { StatsTableHeader } from '@/components/organisms/StatsTableHeader'
+import { useMediaQueryDown } from '@/hooks/style/useMediaQueries'
 import { usePokemonStats } from '@/hooks/usePokemonStats'
 import { useAuthUserState } from '@/store/authUserState'
 import { MAX_EV, MAX_TOTAL_EV } from '@/utils/constants'
-import { useMediaQueryDown } from '@/utils/theme'
 import type { NullableStats, Pokemon, StatsKey } from '@/types/front'
 import type { Nature, PokemonBasicInfo } from '@/types/openapi/schemas'
 import type { PokemonPostParams } from '@/types/openapi/schemas'
