--- src/utils/theme/index.ts
@@ -1,2 +1,0 @@
-export { theme } from './theme'
-export { useMediaQueryUp, useMediaQueryDown } from './mediaQuery'
